reorganize llm code
